requirements.txt
pandas
google-genai
python-dotenv
requests
pypdf
